requirements.txt
absl-py==0.10.0
appdirs==1.4.4
astor==0.8.1
black==20.8b1
certifi==2020.6.20
click==7.1.2
cycler==0.10.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
importlib-metadata==1.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.2
mypy-extensions==0.4.3
numpy==1.19.2
opt-einsum==3.3.0
pathspec==0.8.0
Pillow==7.2.0
protobuf==3.13.0
pyparsing==2.4.7
python-dateutil==2.8.1
regex==2020.7.14
scipy==1.1.0
six==1.15.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
